net: shell: Add pkt command
Move "pkt" command to a separate file.
